added nodemon and socket.io -- changed file structure
