+++ backend/src/app/app.ts
@@ -1,0 +1,13 @@
+import express from "express";
+import { 
+  Request, 
+  Response 
+} from "express"
+
+const app = express();
+
+app.get("/", (req: Request, res: Response) => {
+  res.send("Hello world!");
+});
+
+export default app;
