+++ backend/src/index.ts
@@ -1,0 +1,5 @@
+import app from "./app/app";
+
+app.listen(8000, () => {
+  return console.log('Express is listening at "http://localhost:8000"');
+});
